# Conflicts:
#	build.gradle
#	src/main/java/stackpot/stackpot/domain/PotRecruitmentDetails.java
#	src/main/java/stackpot/stackpot/service/MyPotServiceImpl.java
#	src/main/java/stackpot/stackpot/service/PotServiceImpl.java
#	src/main/java/stackpot/stackpot/web/controller/PotController.java
#	src/main/java/stackpot/stackpot/web/dto/PotMemberResponseDTO.java
#	src/main/resources/application.yml

+++ src/main/java/stackpot/stackpot/converter/FeedConverterImpl.java
@@ -9,7 +9,7 @@ import stackpot.stackpot.web.dto.FeedResponseDto;
 public class FeedConverterImpl implements FeedConverter{
 
     @Override
-    public FeedResponseDto.FeedDto feedDto(Feed feed, int popularity, int likeCount) {
+    public FeedResponseDto.FeedDto feedDto(Feed feed, long popularity, long likeCount) {
         return FeedResponseDto.FeedDto.builder()
                 .id(feed.getFeedId())
                 .writer(feed.getUser().getNickname())
